+++ Packages/ai.fxn.fxn3d/Runtime/Function.cs
@@ -103,7 +103,7 @@ namespace Function {
 
         #region --Operations--
         public readonly FunctionClient client;
-        public const string Version = @"0.0.16";
+        public const string Version = @"0.0.17";
         internal const string URL = @"https://api.fxn.ai";
         #endregion
     }